+++ NEWS.md
@@ -1,6 +1,13 @@
 # amt 0.0.4.0
+
+## major changes
+- Added an alias to `mk_track` called `make_track`.
+- `mk_track` gained a new argument `all_cols` that allows to carry over all columns to track.
+- `mk_track` gained a new argument `check_duplicates` that allows to the user to specify if temporal duplicates should give an error or not.
+
 ## minor changes
 - Removed: Depends tidyverse
+- Fixed errors with lubridate::Periods within `data_frame`s
 
 # amt 0.0.3.0
 ## minor changes