+++ NEWS.md
@@ -1,11 +1,12 @@
 # amt 0.0.4.0
 
 ## major changes
-- Added kenernel home ranges
 - Added an alias to `mk_track` called `make_track`.
 - `mk_track` gained a new argument `all_cols` that allows to carry over all columns to track.
 - `mk_track` gained a new argument `check_duplicates` that allows to the user to specify if temporal duplicates should give an error or not.
 - Added `hr_kde` to calculate Kernel home ranges.
+- `steps` now uses `difftime` to calcualte time differences.
+
 
 ## minor changes
 - Removed: Depends tidyverse